Update issue templates
A issue template standard is needed.  Ideally we should have a best practices for issues that mirrors the RFC. For example, we should adopt the modal language commonly used in technical documentation https://www.rfc-editor.org/rfc/rfc2119.

This is a suggested starting point, but very open to changes/additions, as it is very bare bones.