Generalize getUsdLiquidity 2:
- added two mappings to storage:
twaReserves -> 2 reserves values packed into 1 slot.
`twaReserves` is a mapping from well to twaReserves. holds the twaReserves of a well during sunrise, and returns 1 otherwise.
`usdTokenPrice` is a mapping from well to usdTokenPrice. Holds the usd value of the non-bean token of a well during the sunrise function. returns 1 otherwise.

- added set, reset, and get twaReserves/usdTokenPrice to libWell.
 generalized LibEvaluate to get usd liquidity on all wells, assuming a oracle is added.

- note: moved `updateStalkPerBdvPerSeason` to the gaugePointFacet, as otherwise the seasonFacet would be too large (currently its 23.991kb... should discuss where else we could shave size).