Update PROPOSALS.md
# BOP-4: Formalize BFC Reporting

## Proposer 

Rex (rexthebean.eth)

## Summary

Formalize reporting responsibilities for the BFC:

- Weekly written updates - To be published in the Announcements channel every Monday by 8:00 PM UTC
- Monthly update calls - To be held within six hours of the publication of the first weekly update each month

## Rationale

The Beanstalk DAO depends on Beanstalk developers' information for an understanding of objectives, approaches, and progress. While it’s understood that development is being conducted under an existing budget and that a future development budget may not be necessary (reducing the need for fiduciary accountability), long-term confidence in the protocol is closely tied to DAO members’ understanding of ongoing development and maintenance activities. The BFC is uniquely positioned to provide such updates and, therefore, the appropriate group to hold the responsibility.

## Proposed Solution

Update BFC Responsibilities to include weekly written updates and monthly update calls.

Weekly written updates will include (but need not be limited to):

_Major Projects_
- Name and short description 
- Recent developments (past week)
- Next steps (current week)
- Percentage completion (10% increments) and confidence level in that percentage (P10/P50/P90)

_Ongoing Operational and Maintenance Efforts_
- Tasks completed over the past week (anything that takes more than 60 minutes to complete)
- Tasks for the current week that will likely take more than 60 minutes to complete

_Projects Under Consideration_
- Project description 
- Potential benefit of the project 
- Potential resource necessary for completion 
- Level of consideration (e.g., Gathering Information, Under Review, In Preparation, etc.)

Monthly update calls will include (but need not be limited to):
- A summary of completed project work from the prior month 
- A summary of anticipated project work for the current month 
- A summary of projects under consideration 
- A period for questions and answers from attendees

Monthly update calls will be recorded and made available to the Beanstalk DAO via Discord within 24 hours of completion.

All updates must be performed by a member of the BFC or their appointed representative(s). Failure to deliver updates (written or calls) on more than four occasions in a calendar year will result in the immediate dismissal of all sitting members of the BFC and trigger nominations for new members.

## Beans Minted

None.

## Effective

Upon passage.